Added asset default-config.json parsing

+++ FantasmoSDK/app/src/main/java/com/fantasmo/sdk/utilities/MovingAverage.kt
@@ -2,11 +2,11 @@ package com.fantasmo.sdk.utilities
 
 class MovingAverage(private var period: Int = 30) {
     private var index = 0
-    private var samples: ArrayList<Double> = ArrayList()
+    private var samples: ArrayList<Float> = ArrayList()
 
     var average = (samples.sum()) / (samples.size)
 
-    fun addSample(value: Double): Double {
+    fun addSample(value: Float): Float {
         if (samples.size == period) {
             samples[index] = value
             index = (index + 1) % period
